finish delegate fetch active flow dao
