+++ azkaban-common/src/main/java/azkaban/executor/JdbcExecutorLoader.java
@@ -34,7 +34,6 @@ import java.sql.SQLException;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.apache.commons.dbutils.DbUtils;
@@ -53,6 +52,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   private final ExecutionLogsDao executionLogsDao;
   private final ExecutorEventsDao executorEventsDao;
   private final ActiveExecutingFlowsDao activeExecutingFlowsDao;
+  private final FetchActiveFlowDao fetchActiveFlowDao;
   private EncodingType defaultEncodingType = EncodingType.GZIP;
 
   @Inject
@@ -62,7 +62,8 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
                             final ExecutionJobDao executionJobDao,
                             final ExecutionLogsDao executionLogsDao,
                             final ExecutorEventsDao executorEventsDao,
-                            final ActiveExecutingFlowsDao activeExecutingFlowsDao) {
+                            final ActiveExecutingFlowsDao activeExecutingFlowsDao,
+                            final FetchActiveFlowDao fetchActiveFlowDao) {
     super(props, commonMetrics);
     this.executionFlowDao = executionFlowDao;
     this.executorDao = executorDao;
@@ -70,6 +71,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
     this.executionLogsDao= executionLogsDao;
     this.executorEventsDao = executorEventsDao;
     this.activeExecutingFlowsDao = activeExecutingFlowsDao;
+    this.fetchActiveFlowDao = fetchActiveFlowDao;
   }
 
   public EncodingType getDefaultEncodingType() {
@@ -143,38 +145,15 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   @Override
   public Map<Integer, Pair<ExecutionReference, ExecutableFlow>> fetchActiveFlows()
       throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchActiveExecutableFlows flowHandler = new FetchActiveExecutableFlows();
 
-    try {
-      final Map<Integer, Pair<ExecutionReference, ExecutableFlow>> properties =
-          runner.query(FetchActiveExecutableFlows.FETCH_ACTIVE_EXECUTABLE_FLOW,
-              flowHandler);
-      return properties;
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException("Error fetching active flows", e);
-    }
+    return this.fetchActiveFlowDao.fetchActiveFlows();
   }
 
   @Override
   public Pair<ExecutionReference, ExecutableFlow> fetchActiveFlowByExecId(final int execId)
       throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
-    final FetchActiveExecutableFlowByExecId flowHandler = new FetchActiveExecutableFlowByExecId();
 
-    try {
-      final List<Pair<ExecutionReference, ExecutableFlow>> flows =
-          runner.query(FetchActiveExecutableFlowByExecId.FETCH_ACTIVE_EXECUTABLE_FLOW_BY_EXECID,
-              flowHandler, execId);
-      if(flows.isEmpty()) {
-        return null;
-      }
-      else {
-        return flows.get(0);
-      }
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException("Error fetching active flows by exec id", e);
-    }
+    return this.fetchActiveFlowDao.fetchActiveFlowByExecId(execId);
   }
 
   @Override
@@ -665,130 +644,6 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
                 ExecutableFlow.createExecutableFlowFromObject(flowObj);
 
             execFlows.add(exFlow);
-          } catch (final IOException e) {
-            throw new SQLException("Error retrieving flow data " + id, e);
-          }
-        }
-      } while (rs.next());
-
-      return execFlows;
-    }
-  }
-
-  private static class FetchActiveExecutableFlows implements
-      ResultSetHandler<Map<Integer, Pair<ExecutionReference, ExecutableFlow>>> {
-    // Select running and executor assigned flows
-    private static final String FETCH_ACTIVE_EXECUTABLE_FLOW =
-      "SELECT ex.exec_id exec_id, ex.enc_type enc_type, ex.flow_data flow_data, et.host host, "
-        + "et.port port, et.id executorId, et.active executorStatus"
-        + " FROM execution_flows ex"
-        + " INNER JOIN "
-        + " executors et ON ex.executor_id = et.id"
-        + " Where ex.status NOT IN ("
-        + Status.SUCCEEDED.getNumVal() + ", "
-        + Status.KILLED.getNumVal() + ", "
-        + Status.FAILED.getNumVal() + ")";
-
-    @Override
-    public Map<Integer, Pair<ExecutionReference, ExecutableFlow>> handle(
-        final ResultSet rs) throws SQLException {
-      if (!rs.next()) {
-        return Collections.emptyMap();
-      }
-
-      final Map<Integer, Pair<ExecutionReference, ExecutableFlow>> execFlows =
-          new HashMap<>();
-      do {
-        final int id = rs.getInt(1);
-        final int encodingType = rs.getInt(2);
-        final byte[] data = rs.getBytes(3);
-        final String host = rs.getString(4);
-        final int port = rs.getInt(5);
-        final int executorId = rs.getInt(6);
-        final boolean executorStatus = rs.getBoolean(7);
-
-        if (data == null) {
-          execFlows.put(id, null);
-        } else {
-          final EncodingType encType = EncodingType.fromInteger(encodingType);
-          final Object flowObj;
-          try {
-            // Convoluted way to inflate strings. Should find common package or
-            // helper function.
-            if (encType == EncodingType.GZIP) {
-              // Decompress the sucker.
-              final String jsonString = GZIPUtils.unGzipString(data, "UTF-8");
-              flowObj = JSONUtils.parseJSONFromString(jsonString);
-            } else {
-              final String jsonString = new String(data, "UTF-8");
-              flowObj = JSONUtils.parseJSONFromString(jsonString);
-            }
-
-            final ExecutableFlow exFlow =
-                ExecutableFlow.createExecutableFlowFromObject(flowObj);
-            final Executor executor = new Executor(executorId, host, port, executorStatus);
-            final ExecutionReference ref = new ExecutionReference(id, executor);
-            execFlows.put(id, new Pair<>(ref, exFlow));
-          } catch (final IOException e) {
-            throw new SQLException("Error retrieving flow data " + id, e);
-          }
-        }
-      } while (rs.next());
-
-      return execFlows;
-    }
-  }
-
-  private static class FetchActiveExecutableFlowByExecId implements
-      ResultSetHandler<List<Pair<ExecutionReference, ExecutableFlow>>> {
-    private static final String FETCH_ACTIVE_EXECUTABLE_FLOW_BY_EXECID =
-        "SELECT ex.exec_id exec_id, ex.enc_type enc_type, ex.flow_data flow_data, et.host host, "
-            + "et.port port, et.id executorId, et.active executorStatus"
-            + " FROM execution_flows ex"
-            + " INNER JOIN "
-            + " executors et ON ex.executor_id = et.id"
-            + " Where ex.exec_id = ? AND ex.status NOT IN ("
-            + Status.SUCCEEDED.getNumVal() + ", "
-            + Status.KILLED.getNumVal() + ", "
-            + Status.FAILED.getNumVal() + ")";
-
-    @Override
-    public List<Pair<ExecutionReference, ExecutableFlow>> handle(final ResultSet rs)
-        throws SQLException {
-      if (!rs.next()) {
-        return Collections.emptyList();
-      }
-
-      final List<Pair<ExecutionReference, ExecutableFlow>> execFlows =
-          new ArrayList<>();
-      do {
-        final int id = rs.getInt(1);
-        final int encodingType = rs.getInt(2);
-        final byte[] data = rs.getBytes(3);
-        final String host = rs.getString(4);
-        final int port = rs.getInt(5);
-        final int executorId = rs.getInt(6);
-        final boolean executorStatus = rs.getBoolean(7);
-
-        if (data == null) {
-          logger.error("Found a flow with empty data blob exec_id: " + id);
-        } else {
-          final EncodingType encType = EncodingType.fromInteger(encodingType);
-          final Object flowObj;
-          try {
-            if (encType == EncodingType.GZIP) {
-              final String jsonString = GZIPUtils.unGzipString(data, "UTF-8");
-              flowObj = JSONUtils.parseJSONFromString(jsonString);
-            } else {
-              final String jsonString = new String(data, "UTF-8");
-              flowObj = JSONUtils.parseJSONFromString(jsonString);
-            }
-
-            final ExecutableFlow exFlow =
-                ExecutableFlow.createExecutableFlowFromObject(flowObj);
-            final Executor executor = new Executor(executorId, host, port, executorStatus);
-            final ExecutionReference ref = new ExecutionReference(id, executor);
-            execFlows.add(new Pair<>(ref, exFlow));
           } catch (final IOException e) {
             throw new SQLException("Error retrieving flow data " + id, e);
           }
